Link List
Create an array-based list or a linked-list that allows:

an add function that takes a value and inserts it into a given position into the list
(example: myList.add(someValue, somePosition) )

a remove function that takes a position and removes the value stored at that position of the list and returns it
(example: myList.remove(somePosition) )

a get function that takes a position and returns that value without removing it
(example: myList.get(somePosition) )

Be sure to include at least one test function for each piece of functionality that should verify that your code is working!  This should be at least one test per behavior, likely more.  You can make these tests in a source file with a main where your tests are either directly in the main or inside their own standalone functions (please do not neglect the importance of testing!)

(Note winter 2021: we will cover the analysis of some of this in class next week, then we will have you analyze the next ones!) Analyze the complexity of your implementation with line-by-line analysis

Note: This assignment is to get you to think about the trade-offs that we may have to weigh before using one structure over another